+++ students/dima_lieshchyn/module-js-exam/src/js/dev-modules/getCommentsList.js
@@ -1,3 +1,4 @@
 export default function getCommentsList() {
+
     return localStorage.getItem('userList') ? JSON.parse(localStorage.getItem('userList')) : '';
 }
